+++ src/api/helpers.js
@@ -4,8 +4,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/* global chrome */
-
 import zd from 'zepto-detect'
 import { get, init, last } from 'fkit'
 import { fromEventPattern } from '../lib/rxjs'