+++ src/components/BookmarkForm.svelte
@@ -1,22 +1,12 @@
 <script>
-  import { onDestroy, onMount } from 'svelte'
-  import { bookmarkStore as bookmark } from '../lib/stores'
   import { createBookmark } from '../api/bookmarks'
-  import { messages } from '../lib/messaging'
-  import { t as translate } from '../lib/translate'
-  import BookmarkCount from './BookmarkCount.svelte'
+  import { t } from '../lib/translate'
   import CategorySelector from './CategorySelector.svelte'
   import Favicon from './Favicon.svelte'
 
-  export let t = translate
+  export let bookmark
   let form
   let submitButton
-
-  const popupHeader = () => {
-    return $bookmark.saved
-      ? t('saved_bookmark')
-      : t('add_bookmark')
-  }
 
   const onSubmit = (event) => {
     const valid = form.reportValidity()
@@ -35,57 +25,35 @@
       createBookmark({
         parentId: params.parentId,
         title: subcategory,
-      }, (subcategory) => {
+      }).then((subcategory) => {
         params.parentId = subcategory.id
-        createBookmark(params, () => {
-          $bookmark.saved = true
+        createBookmark(params).then(() => {
+          bookmark.saved = true
         })
       })
     } else {
-      createBookmark(params, () => {
-        $bookmark.saved = true
+      createBookmark(params).then(() => {
+        bookmark.saved = true
       })
     }
 
     window.close()
     return false
   }
-
-  const onTabUpdate = (tab) => {
-    $bookmark = { ...tab }
-  }
-
-  onMount(() => {
-    messages.on('currentTabInfo', onTabUpdate)
-  })
-
-  onDestroy(() => {
-    messages.off('currentTabInfo', onTabUpdate)
-  })
 </script>
 
 <style>
-  :scope {
-    display: block;
-  }
-
-  h1 {
-    color: #333;
-  }
-
   .btn[type=submit] {
     padding: .25rem .8rem
   }
 </style>
-
-<h1>{ popupHeader() } <BookmarkCount /></h1>
 
 <form bind:this={form} on:submit|preventDefault={onSubmit}>
   <div class="form-group">
     <label for="title">{ t('title') }</label>
     <input name="title" required
            class="form-input"
-           bind:value={$bookmark.title}>
+           bind:value={bookmark.title}>
   </div>
 
   <div class="form-group">
@@ -93,8 +61,8 @@
     <div class="input-group">
       <input name="url" type="url" required
              class="form-input"
-             bind:value={$bookmark.url}>
-      <Favicon icon={$bookmark.favIconUrl}></Favicon>
+             bind:value={bookmark.url}>
+      <Favicon icon={bookmark.favIconUrl}></Favicon>
     </div>
   </div>