+++ CHANGELOG.md
@@ -7,6 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.2] – 2021-05 Bug fixes and stability
+
+### Fixed
+
+- 7809216 Fix popup header to show when a bookmark is saved
+- bc1243b Fix using subcategory close button with keyboard
+- 022ee1f Fix close button on popup
+
+### Added
+
+- 913bbb3 Show bookmark count on bookmark form
+- 9df8570 Show badge when several bookmarks exist
+
+### Changed
+
+- 2e9edb6 Change the minimum length of search back to two letters
+
+**Technical:**
+
+- 1d8e9d8 rollup: Use absolute paths to prevent erronous destinations on copy
+- 349cb6c Switch from node-sass to sass
+- 2274c3b Get recent categories using getRecent API method
+- 36cedaa Simplify message handling with generic messageBridge
+- 702dd06 Start using pnpm instead of yarn
+- 01e2b64 Replace @riotjs/observable with browserify/events
+
 # Released versions:
 
 ## [0.4.1] – 2019-12-22 – Fix bookmark paths